Explicit cast to (size_t) in ccache_read()
This is safe because:
- fpos is positive (we already checked that the result of fpos = ftello(R.f);
  was not -1).
- pagesize = sysconf(_SC_PAGESIZE); is definitely going to fit into a (size_t).
- therefore (fpos % pagesize) will also fit into a (size_t).

+++ tar/ccache/ccache_read.c
@@ -287,7 +287,7 @@ ccache_read(const char * path)
 		    (intmax_t)(sb.st_size), (intmax_t)(fpos + R.datalen));
 		goto err5;
 	}
-	C->datalen = R.datalen + (fpos % pagesize);
+	C->datalen = R.datalen + (size_t)(fpos % pagesize);
 	if ((C->data = mmap(NULL, C->datalen, PROT_READ,
 #ifdef MAP_NOCORE
 	    MAP_PRIVATE | MAP_NOCORE,
